feat(games): lazy-load non-game surfaces, normalize routes, add READMEs
All trailer/editor/viewer/gallery/tune/bake surfaces now load via dynamic
import() only when their param is present — zero surface code in main
chunks (verified per-game via build output). Route convention: ?trailer
?editor ?viewer ?gallery ?tune ?bake, query params only, presence-checked.
Renames: farm #gallery -> ?gallery, lunerfall ?editor -> ?viewer (scene
file/class renamed to match), moba ?ui -> ?viewer + ?gallery=<section>.
crazy-waymo: tune-panel + bake machinery extracted to lazy chunks
(world-bin split read/write). READMEs in all 11 games with route tables.
